+++ alert_generator/testsuite/cases/interface.go
